Created Main and Config.json
Created python file main to act as the main method for the project. Created config.json in order to hold the configuration data for runtime. Main method pulls the data from the config file and then as of right now is only capable of pulling the plex movie library data and saving it into a csv file. More functionality planned for future versions.