Create user "database". Route to add users
There is now a users.json file within the data folder that holds the json for all of the created users. The structure is set up so that each object is a user. The key is their username so to ensure that each user is unique. The value is the full user object.